Added bracket formation and error checking
Added feature to generate the correct amount of empty matches, separated
in to rounds to use on tournament creation. Tournament class will be
used to generate tournaments and store multiple schedules, one for each
round. Created InvalidNumTeamsException to be thrown when an invalid
amount of teams (not divisible down to a final match) are passed to a
Tournament.